requirements.txt
allure-pytest>=2.13.5
pytest>=8.2.1
selene>=2.0.0rc9